+++ src/Pim/Bundle/DataGridBundle/Resources/public/js/datagrid/action/configure-columns-action.js
@@ -154,7 +154,7 @@ define(
                         if (values.length) {
                             DatagridState.set(this.gridName, 'columns', values.join(','));
                             var url = window.location.hash;
-                            Backbone.history.navigate(url.substr(0, url.length -1));
+                            Backbone.history.fragment = new Date().getTime();
                             Backbone.history.navigate(url, true);
                         } else {
                             messenger.notificationFlashMessage('error', __('datagrid_view.columns.min_message'));