+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Bug fixes
 - PIM-3786: Attribute type should not be blank for import
+- PIM-3437: Fix applying datagrid views and columns when not using hash navigation
 
 # 1.2.31 (2015-03-06)